pyproject.toml
[project]
name = "JasmineTool"
version = "0.1.0"
description = "Automated multi-GPU/multi-host orchestration via SSH"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=5.4.0",
    "setuptools>=75.3.2",
    "toml>=0.10.2",
]
keywords = ["gpu", "ssh", "orchestration", "parallel", "wandb", "sweep", "distributed"]
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Distributed Computing",
    "Topic :: System :: Systems Administration",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/JasmineTool"
"Bug Reports" = "https://github.com/yourusername/JasmineTool/issues"
"Source" = "https://github.com/yourusername/JasmineTool"

[project.scripts]
jasminetool = "jasminetool.cli:main"
jasmine-tool = "jasminetool.cli:main"
jt = "jasminetool.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.812",
]
docs = [
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
